Goal: attempting to start simulation with data loaded to ECS instead of old system;
Problem: orbit loads and looks for a preloaded system, but ECS hasn't initialized yet.

When I either
	A. prevent orbit from preloading meshes at the beginning or
	B. initialize ECS before scene.orbit
I get an invisible ship, but the orbit (and the planet) renders fine!

Hypothesis: there is an interaction/relationship between data for orbital mesh and ship mesh;
Actionable steps: investigate how sys was populated before, and how using ECS to populate it alters initialization of ship meshes? Where are the ship meshes being loaded? Does it depend on the existence of the orbital meshes?